requirements.txt
runpod==1.7.0
diffusers==0.31.0
transformers==4.46.0
accelerate==1.1.0
safetensors==0.4.5
torch==2.4.0
Pillow==10.4.0
sentencepiece==0.2.0
protobuf==5.28.0
huggingface-hub==0.26.0
optimum-quanto==0.2.4
